+++ Runtime/PoolModule/Data/ValueObjects/PoolItemBaseCVO.cs
@@ -6,10 +6,8 @@ namespace FlowIoC.PoolModule.Data.ValueObjects
     [Serializable]
     public abstract class PoolItemBaseCVO
     {
-        [Header("Identification")]
         public string PoolKey;
 
-        [Header("Pool Config")]
         public int InitialCreateCount = 10;
         public bool IsExtendable = true;
         public bool LazyLoad = false;